Extracted utilities into their own classes.
* Moved existing helper code into appropriate classes.  This commit
  includes no new code, but comments were added.
* Added util package.
* Added ArrayIterator.java and IterableArray.java providing for arrays
  as iterables.
* Added Arrays.java containing array utilities.  So far, this includes
  conversions from primitive arrays to object arrays.
* Added Assert.java containing assertion testing utilities.  So far,
  this includes comparing iterables.
* Added 'allclean' make target.
